+++ src/SFA.DAS.Notifications.Infrastructure.UnitTests/SFA.DAS.Notifications.Infrastructure.UnitTests.csproj
@@ -62,7 +62,8 @@
     <Compile Include="ConfigurationTests\TemplateConfigurationServiceTests\TemplateConfigurationServiceTestBase.cs" />
     <Compile Include="ConfigurationTests\TemplateConfigurationServiceTests\WhenGettingAsync.cs" />
     <Compile Include="ConfigurationTests\TemplateConfigurationServiceTests\WhenGetting.cs" />
-    <Compile Include="NotifyEmailServiceTests\NotifyEmailServiceTests\WhenSending.cs" />
+    <Compile Include="NotifyEmailServiceTests\WhenSendingAnSms.cs" />
+    <Compile Include="NotifyEmailServiceTests\WhenSendingAnEmail.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -86,6 +87,7 @@
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
